requirements.dev.txt
django-admin >= 2.0.1, < 3.0.0
django-admin-interface >= 0.19.1, < 1.0.0
psycopg2          >= 2.9.3, < 3.0.0
# psycopg2-binary >= 2.9.3, < 3.0.0
flake8
black
